requirements.txt
matrix-nio
nio-bot
pyyaml
requests